ath79: GL-AR300M: Provide NAND support
The GL.iNet GL-AR300M has been supported by the ar71xx and ath79
platforms with access to its 16 MB NOR flash, but not its 128 MB
SPI NAND flash.

This commit provides support for the NAND through the upstream
SPI-NAND framework. Devices with both NOR and NAND flash can support
independent firmware on each, with U-Boot able to boot from either.
The OEM U-Boot will fall back to the NOR firmware after three
"unsuccessful" boots.

The family of GL-AR300M devices on the ath79 platform now includes:

  * glinet,gl-ar300m-lite	"generic" target, NOR-only board
  * glinet,gl-ar300m-nand	"nand" target
  * glinet,gl-ar300m-nor	"nand" target (NAND-aware)

Existing OpenWrt and GL.iNet firmware can be upgraded from NOR
versions to glinet,gl-ar300m-nor and from NAND versions to
glinet,gl-ar300m-nand without --force:

  (NOR)  gl-ar300m ==> glinet,gl-ar300m-nor ==> glinet,gl-ar300m-nand
  (NAND) gl-ar300m ==> glinet,gl-ar300m-nand

  glinet,gl-ar300m-nor <=> glinet,gl-ar300m-nand

Upgrades are not possible using sysupgrade from legacy NOR to NAND

The GL-AR300M OEM U-Boot allows upload and flashing of either NOR
firmware (sysupgrade.bin) or NAND firmware (factory.img) through its
HTTP-based GUI. Serial connectivity is not required.

The boot counter is implemented by the OEM using u-boot-env. At this
time, it does not appear that the switch on the side of the unit can
be used to select NOR vs. NAND boot and the fail-over is only from
NAND to NOR. To save flash wear, it is only reset when running the
glinet,gl-ar300m-nand firmware.

NAND-specific base-files are used to remove impact on existing
generic and tiny targets.

As there is now no "generic" build appropriate for the GL-AR300M16,
(or for users of the GL-AR300M that do not need access to NAND)
it will be introduced in a subsequent commit.

The ar71xx targets are unmodified.

Signed-off-by: Jeff Kletsky <git-commits@allycomm.com>