build: sysupgrade-tar alt-board= for legacy upgrades
Targets that use nand_do_platform_check() can't use SUPPORTED_DEVICES
as the check requires ./sysupgrade-legacy_boardname/CONTROL to be
non-zero length as extracted from the tar file. Previously, only
./sysupgrade-new_boardname/CONTROL was present.

This prevents upgrade without --force from, for example, ar71xx to ath79

Add an optional alt-board= parameter to parsing of sysupgrade-tar that
creates a directory at the expected location for the alt-board name,
copying over the CONTROL file.

The contents of CONTROL are unmodified by this commit.

Careful ordering of the tar file is required as existing builds
expect the kernel and root assets to be in the first directory
returned by tar -tf that matches sysupgrade-*/

Run-tested-on: EA8300, GL-AR300M, GL-AR750S

Signed-off-by: Jeff Kletsky <git-commits@allycomm.com>