ramips/mt7621: add mt7530 port 5 support.
This patch add support for port 5 in a similar way as upstream kernel [0].
Because we don't have DSA to parse the port setup, I added some code to
detect the port setup and phy-mode. Setup port 5 based on the detection.

The port 5 setup code supports also that port 0 of port 4 can be connected
to 2nd GMAC but the rest of the internals doesn't support that!

[0]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?h=v5.4-rc3&id=38f790a805609b255f0cacc097e2df4324a3d041
Signed-off-by: René van Dorst <opensource@vdorst.com>