ramips/mt7621: ubiquiti_edgerouterx-sfp: Enable i2c-clk gate for SFP cage
SFP cage is also connected to i2c of the SOC.
But the i2c clock to the SFP cage is gated.
Enable the gate to able to read the SFP module.

Gpio-led is currently a hack. I think the right solution is gpio-hogs.

Use i2c-tools to readout the sfp module.

like: i2cdump -y 0 0x50

Signed-off-by: René van Dorst <opensource@vdorst.com>