ramips: mt7620: allow both internal and external EPHYs
Now that MDIO mode properly sets the default EPHY base address to 12,
and that the EPHY base address can be set to a custom value if needed,
it is no longer necessary to disable internal PHYs or have an else block for MDIO mode.

In theory, this allows for boards that have both external switches
and internal PHYs that lead to ethernet ports to be supported.

Therefore, move the lines to the function main block,
and have only one instance of the function mtk_switch_w32
for writing the register for EPHY base address.

Signed-off-by: Michael Pratt <mcpratt@pm.me>